+++ src/index.tsx
@@ -4,7 +4,7 @@ export { default as Badge } from 'badge'
 export { default as Button, ButtonStatus } from 'button'
 export { default as ButtonGroup } from 'buttonGroup'
 export { default as Caption } from 'caption'
-export { default as ConfirmationModal } from 'confirmationModal'
+export { default as ConfirmationModal, ConfirmationModalStatus } from 'confirmationModal'
 export { default as DatePicker, DatePickerOrientation } from 'datePicker'
 export { default as Drawer } from 'drawer'
 export { default as DropdownButton } from 'dropdownButton'
